+++ app/scripts/lib/lforms-data.js
@@ -684,7 +684,6 @@
 
     /**
      * Validate user input value
-     * Note: Not currently used since validations are handled in an Angular directive.
      * This might be used in the future.
      * @param item the question item
      * @private
@@ -1504,7 +1503,7 @@
         }
       }
 
-      var defData = {
+      var defData = { 
         lformsVersion: this.lformsVersion,
         PATH_DELIMITER: this.PATH_DELIMITER,
         code: this.code,
@@ -1556,6 +1555,34 @@
       }
       // return a deep copy of the data
       return angular.copy(ret);
+    },
+
+
+    /**
+     * Get a list of errors preventing the form from being valid.
+     * @returns {Array<string> | null} list of errors or null if no errors
+     */
+    checkValidity: function () {
+      const errors = [];
+      const itemListLength = this.itemList.length;
+
+      for (let i = 0; i < itemListLength; i++) {
+        const item = this.itemList[i];
+
+        this._checkValidations(item);
+
+        if (item._validationErrors !== undefined && item._validationErrors.length) {
+          const errorDetails = item._validationErrors.map((e) => `${item.question} ${e}`);
+
+          Array.prototype.push.apply(errors, errorDetails);
+        }
+      }
+
+      if (errors.length) {
+        return errors;
+      } else {
+        return null;
+      }
     },