+++ CHANGELOG.md
@@ -3,9 +3,15 @@
 This project follows [Semantic Versioning](http://semver.org/).
 
 
-## [26.2.1] 2020-10-19
+## [26.3.1] 2020-10-19
 ### Fixed
 - Fixed a bug in STU3 Questionnaire import that 'options' is not handled correctly
+
+## [26.3.0] 2020-10-16
+### Added
+- Added two functions: LForms.Util.isValid, which checks if the form fields are
+  valid and LForms.Util.getErrors to get a list of errors preventing the form
+  from being valid.
 
 ## [26.2.0] 2020-10-06
 ### Added